I.A/Assist/NutriAI/requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
pydantic
Pillow
flask_cors
flask